WIP Graphviz impl

+++ neural/src/math.rs
@@ -1,5 +1,5 @@
 
-pub type Float = f32;
+pub type Float = f64;
 
 #[macro_export]
 macro_rules! forw_calc {
@@ -11,6 +11,6 @@ macro_rules! forw_calc {
 #[macro_export]
 macro_rules! sigmoid {
     ($a:expr) => (
-        1.0/1.0+Float::powf(std::f32::consts::E,-$a)
+        1.0/(1.0+Float::powf(std::f32::consts::E as Float,-$a))
     );
 }
